📦 Remove TracingContext and use strings instead
The TracingContext enum was just adding an extra step to every begin and end context where the string had to be looked up, and it was never actually used other than finding those strings. Instead, pass the context type as a string to the tracing method.

To get around the issue of renaming `name=“”` for rulesets, add a new parameter to traceBeginLabelledContext() giving the name of the attribute, and add another helper function, traceBeginNamedContext(), which defaults that attribute to “name”.