🐛 Fix a subtle mistake in the tracing of fail
If fail is called in the condition of if/try, or in a loop, the entire program does not terminate. In fact, it simply just exits that context. This means that when fail is used, we have to exit contexts until we see a <condition>, in the case of branches.
For loops, we do not need to exit any contexts, because the loop does not exit early at runtime.